Add project_manage_access check in Projects project editing views and projects.html

+++ projects/views.py
@@ -45,43 +45,46 @@ class Projects:
     @staticmethod
     @login_required
     def add_project(request):
-        if request.method.lower() == 'post':
-            name = request.POST.get('name')
-            client = Client.objects.get(company=request.user.company, name=request.POST.get('client'))
-            notes = request.POST.get('notes')
-            budget = request.POST.get('budget')
-            new_project = Project.objects.create(
-                company=request.user.company,
-                name=name,
-                client=client,
-                notes=notes,
-                budget=budget
-            )
-            tasks = request.POST.getlist('tasks')
-            for task in tasks:
-                new_project.tasks.add(Task.objects.get(company=request.user.company, name=task))
-            new_project.save()
+        if request.user.role.project_manage_access:
+            if request.method.lower() == 'post':
+                name = request.POST.get('name')
+                client = Client.objects.get(company=request.user.company, name=request.POST.get('client'))
+                notes = request.POST.get('notes')
+                budget = request.POST.get('budget')
+                new_project = Project.objects.create(
+                    company=request.user.company,
+                    name=name,
+                    client=client,
+                    notes=notes,
+                    budget=budget
+                )
+                tasks = request.POST.getlist('tasks')
+                for task in tasks:
+                    new_project.tasks.add(Task.objects.get(company=request.user.company, name=task))
+                new_project.save()
         return redirect('projects')
 
     @staticmethod
     @login_required
     def edit_project(request, pk):
-        if request.method.lower() == 'post':
-            project = Project.objects.get(company=request.user.company, pk=pk)
-            project.name = request.POST.get('name')
-            project.client = Client.objects.get(company=request.user.company, name=request.POST.get('client'))
-            project.notes = request.POST.get('notes')
-            project.budget = request.POST.get('budget')
-            project.tasks.clear()
-            tasks = request.POST.getlist('tasks')
-            for task in tasks:
-                project.tasks.add(Task.objects.get(company=request.user.company, name=task))
-            project.save()
+        if request.user.role.project_manage_access:
+            if request.method.lower() == 'post':
+                project = Project.objects.get(company=request.user.company, pk=pk)
+                project.name = request.POST.get('name')
+                project.client = Client.objects.get(company=request.user.company, name=request.POST.get('client'))
+                project.notes = request.POST.get('notes')
+                project.budget = request.POST.get('budget')
+                project.tasks.clear()
+                tasks = request.POST.getlist('tasks')
+                for task in tasks:
+                    project.tasks.add(Task.objects.get(company=request.user.company, name=task))
+                project.save()
         return redirect('projects')
 
     @staticmethod
     @login_required
     def delete_project(request, pk):
-        project = Project.objects.get(company=request.user.company, pk=pk)
-        project.delete()
+        if request.user.role.project_manage_access:
+            project = Project.objects.get(company=request.user.company, pk=pk)
+            project.delete()
         return redirect('projects')
